requirements.txt
gradio
ultralytics
numpy
pillow
